Move IsOptional to the internal module.
