+++ fresnel/src/Fresnel/Optional.hs
@@ -20,13 +20,13 @@ module Fresnel.Optional
 , unpackedOptional
 ) where
 
-import Control.Arrow (Kleisli)
 import Data.Bifunctor
 import Data.Maybe (isJust, isNothing)
 import Data.Profunctor
 import Fresnel.Iso.Internal (IsIso)
 import Fresnel.Lens.Internal (IsLens)
 import Fresnel.Optic
+import Fresnel.Optional.Internal (IsOptional)
 import Fresnel.Prism.Internal (IsPrism)
 import Fresnel.Profunctor.OptionalStar
 
@@ -35,14 +35,6 @@ import Fresnel.Profunctor.OptionalStar
 type Optional s t a b = forall p . IsOptional p => Optic p s t a b
 
 type Optional' s a = Optional s s a a
-
-class (IsLens p, IsPrism p) => IsOptional p where
-
-instance IsOptional (->)
-instance Monad m => IsOptional (Kleisli m)
-instance Monoid r => IsOptional (Forget r)
-instance Applicative f => IsOptional (Star f)
-instance Functor f => IsOptional (OptionalStar f)
 
 
 -- Construction
